Add files via upload
Each file as a purpose. 'DOLA' is the name of the bank. It stands for 'Dollas Bank'. 'DOLA' make the files identifiable. 'DOLA_arduinoThermalPrinter.ino' is used for the arduino that controls the Thermal Printer that is responsible for printing the reciepts. 'DOLA_bankserver.js' is the Node.js server that is used for our bank. 'DOLA_landserver.js' is the Node.js server that is used by all bank within our land, Venezuela. 'DOLA_dispenser.ino' is used for the arduino that controls the motor's that dispense the money. 'DOLA.py' is our libary with functions, that can be called to do all sorts of things. 'DOLA_talen.py' has all text that is displayed in the GUI in different languages. 'DOLA_main.py' runs the GUI and calls the functions in 'DOLA.py' and 'DOLA_talen.py' to make the ATM work.